remove dead code, organize code more
