+++ npm/react/src/index.ts
@@ -1,3 +1,7 @@
-export * from './mount'
+export * from './createMount'
 
 export * from './mountHook'
+
+export * from './mountLegacy'
+
+export * from './types'
